--- swiftify-frontend/src/view/LoadSongsView1.form
@@ -1,101 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="view.LoadSongsView1">
-  <grid id="27dc6" binding="LoadSongsViewPanel" layout-manager="GridLayoutManager" row-count="1" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-    <margin top="0" left="0" bottom="0" right="0"/>
-    <constraints>
-      <xy x="20" y="20" width="500" height="400"/>
-    </constraints>
-    <properties/>
-    <border type="none"/>
-    <children>
-      <splitpane id="dfc2e" binding="MainPanel">
-        <constraints>
-          <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false">
-            <preferred-size width="200" height="200"/>
-          </grid>
-        </constraints>
-        <properties/>
-        <border type="none"/>
-        <children>
-          <scrollpane id="c344f" binding="LeftPanel">
-            <constraints>
-              <splitpane position="left"/>
-            </constraints>
-            <properties/>
-            <border type="none"/>
-            <children>
-              <grid id="7f2a4" binding="SongListPanel" layout-manager="BorderLayout" hgap="0" vgap="0">
-                <constraints/>
-                <properties/>
-                <border type="none"/>
-                <children/>
-              </grid>
-            </children>
-          </scrollpane>
-          <grid id="f1058" binding="RightPanel" layout-manager="GridLayoutManager" row-count="4" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-            <margin top="0" left="0" bottom="0" right="0"/>
-            <constraints>
-              <splitpane position="right"/>
-            </constraints>
-            <properties/>
-            <border type="none"/>
-            <children>
-              <component id="56a4f" class="javax.swing.JProgressBar" binding="SongProgressBar">
-                <constraints>
-                  <grid row="2" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-                </constraints>
-                <properties/>
-              </component>
-              <vspacer id="cefc1">
-                <constraints>
-                  <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="6" hsize-policy="1" anchor="0" fill="2" indent="0" use-parent-layout="false"/>
-                </constraints>
-              </vspacer>
-              <component id="3db24" class="javax.swing.JLabel" binding="CurrentSongField">
-                <constraints>
-                  <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="0" fill="0" indent="0" use-parent-layout="false"/>
-                </constraints>
-                <properties>
-                  <text value="Currently Playing: ..."/>
-                </properties>
-              </component>
-              <grid id="4a202" binding="ControlsPanel" layout-manager="FlowLayout" hgap="5" vgap="5" flow-align="1">
-                <constraints>
-                  <grid row="3" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
-                </constraints>
-                <properties/>
-                <border type="none"/>
-                <children>
-                  <component id="c3a0d" class="javax.swing.JButton" binding="PreviousSongButton">
-                    <constraints/>
-                    <properties>
-                      <text value="&lt;"/>
-                    </properties>
-                  </component>
-                  <component id="556fa" class="javax.swing.JButton" binding="PlayPauseButton">
-                    <constraints/>
-                    <properties>
-                      <text value="||"/>
-                    </properties>
-                  </component>
-                  <component id="77f18" class="javax.swing.JButton" binding="NextSongButton">
-                    <constraints/>
-                    <properties>
-                      <text value="&gt;"/>
-                    </properties>
-                  </component>
-                  <component id="e7316" class="javax.swing.JButton" binding="ShuffleButton">
-                    <constraints/>
-                    <properties>
-                      <text value="?"/>
-                    </properties>
-                  </component>
-                </children>
-              </grid>
-            </children>
-          </grid>
-        </children>
-      </splitpane>
-    </children>
-  </grid>
-</form>
